+++ src/main/java/com/korit/carecheckkoreait/controller/OrderController.java
@@ -1,6 +1,5 @@
 package com.korit.carecheckkoreait.controller;
 
-import com.korit.carecheckkoreait.CareCheckKoreaItBackApplication;
 import com.korit.carecheckkoreait.dto.request.ReqAddOrderDto;
 import com.korit.carecheckkoreait.dto.request.ReqSearchOrderDto;
 import com.korit.carecheckkoreait.service.OrderService;
@@ -10,14 +9,14 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/order")
+@RequestMapping("/orders")
 public class OrderController {
 
     @Autowired
     private OrderService orderService;
 
     @Operation(summary = "오더등록", description = "오더등록")
-    @PostMapping("/enroll")
+    @PostMapping("")
     public ResponseEntity<?> orderEnroll(@RequestBody ReqAddOrderDto reqAddOrderDto) {
         orderService.save(reqAddOrderDto);
         return ResponseEntity.ok().build();
@@ -27,13 +26,5 @@ public class OrderController {
     @GetMapping("/list")
     public ResponseEntity<?> searchOrder(@ModelAttribute ReqSearchOrderDto reqSearchOrderDto) throws Exception {
         return ResponseEntity.ok().body(orderService.getAllOrders(reqSearchOrderDto));
-    }
-
-    @Operation(summary = "점수단가 변경", description = "점수단가 변경")
-    @PostMapping("/score")
-    public ResponseEntity<?> updateScorePay(@RequestParam double scorePay) {
-        System.out.println("controller" + scorePay);
-        orderService.updateScorePay(scorePay);
-        return ResponseEntity.ok().build();
     }
 }
